+++ src/libs/Message.ts
@@ -326,6 +326,12 @@ export default class Message implements Serializable<SerializedMessage> {
 
         parsedMessage[startIndex] = `<span class="mention self">${withAtSign ? '@' : ''}${currentUsername}</span>`
       } else if (word.text === '@') {
+        const previousWord = index > 0 ? words[index - 1] : null
+
+        if (!_.isNil(previousWord) && previousWord.text.trim().length > 0 && previousWord.text !== '') {
+          return
+        }
+
         const nextWord = index < words.length - 1 ? words[index + 1] : null
 
         if (!_.isNil(nextWord) && nextWord.text.trim().length > 0) {